snugh/requirements.txt
asgiref==3.2.10
Django==3.1.12
django-debug-toolbar==2.2.1
django-dotenv==1.4.2
djangorestframework==3.11.2
django-cors-headers==3.7.0
mysqlclient==2.0.1
pytz==2020.1
sqlparse==0.3.1
Pillow==8.3.2
uwsgi
